New fitting user example.
